Traning-app/requirements.txt
datasets==4.1.1
peft==0.17.1
PyYAML==6.0.3
transformers==4.57.1
